Move shouldTargetWpcom to wpcom-proxy-request
