+++ packages/data-stores/package.json
@@ -34,7 +34,6 @@
 		"@automattic/domain-utils": "workspace:^",
 		"@automattic/format-currency": "workspace:^",
 		"@automattic/happychat-connection": "workspace:^",
-		"@automattic/help-center": "workspace:^",
 		"@wordpress/api-fetch": "^6.6.0",
 		"@wordpress/data-controls": "^2.9.0",
 		"@wordpress/deprecated": "^3.9.0",
