+++ packages/help-center/src/hooks/use-help-search-query.ts
@@ -1,8 +1,7 @@
 import apiFetch from '@wordpress/api-fetch';
 import { useQueryClient, useQuery } from 'react-query';
-import wpcomRequest from 'wpcom-proxy-request';
+import wpcomRequest, { shouldTargetWpcom } from 'wpcom-proxy-request';
 import { SearchResult } from '../types';
-import { shouldTargetWpcom } from '../utils';
 
 interface APIFetchOptions {
 	global: boolean;
